Parse input straight into an output handle.
Failure to parse is propogated by exceptions.

When parsing to stdout, a temporary file is used to collect the parsed
data. On failure the exception causes the temporary file to be deleted
without being written to stdout. On success, the temporary file contents
are read and copied to stdout.

When parsing to an output file, a parse exception results in the output
file being deleted. On success, the file remains.